fix 'No enumerators available' - on Windows 10
Fixes #30

dc_full is a pointer, if interpretted as a signed number may be negative. Null -> 0 indicates failure.
Set return types for foriegn function so return value can be compared to
None.

ReleaseDC was not being called with enough arguments, it should also
take the same Window Handle (HWND) as GetDC.